Fix issue where "go fmt" would fail on some paths.
"go fmt dir" works correctly when dir is a relative path like "./out",
but it fails when dir is a relative path without dot or two dots, e.g.,
"out". This happens because "go fmt out" is interpreted as a request to
format package "out" rather than relative directory "out".
To fix the problem, use "gofmt" binary with "-w" flag, it should work
for both "./out" and "out" ways of specifying a relative directory,
since it takes an argument of path rather than Go package.
Fixes #60.