Improve error readability by using Fatalln over Fatal.
Before this fix, errors were less readable because there was no space
after the colon:
    error generating package:gofmt error
After this change, errors are more readable:
    error generating package: gofmt error
Prefer to use Fatalln with a string that ends without a space, since
Fatalln space-separates all arguments. An alternative is to use Fatal
but add a space after colon, but that seems less clean and less common.