allow non null-terminated strings in gl.Strs
This change allows non null-terminated strings to be passed as parameters
to gl.Strs. The original motivation for not allowing such was that it was
not useful, however we have recently discovered that gl.ShaderSource and
probably a few other functions take a `lengths` parameter whereby they do
not strictly need a null-terminated string.

Since go-gl/gl and go-gl/glow are low-level libraries that mimic the C
API, we relax this restriction and document the behavior users can expect.