Panic on invalid input to gl.Ptr.
Currently, gl.Ptr will simply not set value of adds when provided with
a pointer to a type that is not supported. That means it returns null
pointer, which is likely going to cause issues for user of gl.Ptr.
Instead of silently ignoring this, panic with a helpful error message.

Also follow up to #55 and remove mention of arrays, they are not
supported.

Fix for go-gl/gl#32.