Fix static error when converting uintptr in gl.Ptr (#111)
reflect will always panic on this conversion.  As uintptr is an integer type, it is not interpreted as a pointer in itself.  Also, this call is redundant per se as it returns a uintptr.  Just converting data to a uintptr and passing to unsafe.Pointer should get the expected result.